backend/requirements.txt
fastapi==0.120.1
uvicorn==0.38.0
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.2.1
psycopg2-binary==2.9.11
redis==7.0.1
sqlalchemy==2.0.44
alembic==1.17.0
pytest==8.4.2
pytest-asyncio==1.2.0
httpx==0.28.1
# Video processing dependencies
opencv-python==4.10.0.84
Pillow==10.4.0
# Cloud storage dependencies
boto3==1.35.89
google-cloud-storage==2.15.0
# Additional utilities
aiofiles==24.1.0
email-validator==2.2.0
